+++ Snabble/Core/Cart/CheckoutData.swift
@@ -178,6 +178,7 @@ public struct CheckoutProcess: Decodable {
     public let modified: Bool
     public let paymentInformation: PaymentInformation?
     public let paymentState: PaymentState
+    public let orderID: String?
 
     public struct ProcessLinks: Decodable {
         public let `self`: Link