+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # v0.10.2
 
-* Adds support for additional payment methods, `creditCardVisa` and `creditCardMastercard`.
+* Adds support for new payment methods, `creditCardVisa` and `creditCardMastercard`.
 * Adds pull-to-refresh to `ReceiptsListViewController`. 
 * `ReceiptsListViewController.init` now takes an optional `CheckoutProcess` parameter. If this is used, the VC will add the specified order to the table display, if it's not already in the client's order list, and wait for the receipt PDF to be generated.